refactor: move policy action resolution from @c15t/react to c15t core
Extracts the pure policy action utilities (resolvePolicyAllowedActions,
resolvePolicyActionGroups, etc.) from the React package to c15t core.
These are zero-dep functions that resolve backend policy hints into
button layouts — used by both @c15t/react and @c15t/embed.

The React package re-exports from c15t for backward compatibility.
